pyproject.toml
[tool.poetry]
name = "oc-lettings-fr"
version = "0.1.0"
description = "School project - Orange County Lettings Reworking"
authors = ["DayriseA"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10" # Django 5.0.6 requires Python >=3.10
django = "^5.0.6"
psycopg = {extras = ["binary"], version = "^3.1.19"}
python-dotenv = "^1.0.1"
sentry-sdk = {extras = ["django"], version = "^2.2.0"}

[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
pytest-django = "^4.8.0"
coverage = "^7.5.1"
pytest-cov = "^5.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.coverage.run]
omit = [
    "*/migrations/*",
    "*/tests/*",
    "*/__init__.py",
    "*/wsgi.py",
    "*/settings.py",
    "*/asgi.py",
    "manage.py",
]